Added support for separate colours on each zone.
Needs a lot of cleanup and can't save anything other than "all" yet.

Also fixed the checksum, just by cooincidence if all zones were set to the same colour then the checksum was always 0x1E (which I'd sniffed previously), but as soon as one zone is different the checksum changes. Now calculates the checksum instead of hardcoding it.